Fixes: Loan manager Dashboard 2
